requirements/requirements-oracle.txt
Django>=1.4,<1.5
simplejson
oauth2
django-compressor
django-mobility
django-templatetag-handlebars
oauth_provider
mock==1.0.1

-e git+git://github.com/sbutler/django-shibboleth-remoteuser.git#egg=django-shibboleth-remoteuser

#cx_Oracle

